[tagger] Move proto utils to tagger dir
