+++ comp/core/tagger/proto/proto.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-2020 Datadog, Inc.
 
-// Package proto contains protobuf related helpers.
+// Package proto provides conversions between Tagger types and protobuf.
 package proto
 
 import (
