+++ comp/core/tagger/taggerimpl/server/server.go
@@ -15,11 +15,11 @@ import (
 	"google.golang.org/grpc/status"
 
 	"github.com/DataDog/datadog-agent/comp/core/tagger"
+	"github.com/DataDog/datadog-agent/comp/core/tagger/proto"
 	"github.com/DataDog/datadog-agent/comp/core/tagger/taggerimpl/telemetry"
 	pb "github.com/DataDog/datadog-agent/pkg/proto/pbgo/core"
 	"github.com/DataDog/datadog-agent/pkg/util/grpc"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
-	pbutils "github.com/DataDog/datadog-agent/pkg/util/proto"
 )
 
 const (
@@ -43,7 +43,7 @@ func NewServer(t tagger.Component) *Server {
 // and streams them to clients as pb.StreamTagsResponse events. Filtering is as
 // of yet not implemented.
 func (s *Server) TaggerStreamEntities(in *pb.StreamTagsRequest, out pb.AgentSecure_TaggerStreamEntitiesServer) error {
-	cardinality, err := pbutils.Pb2TaggerCardinality(in.Cardinality)
+	cardinality, err := proto.Pb2TaggerCardinality(in.Cardinality)
 	if err != nil {
 		return err
 	}
@@ -70,7 +70,7 @@ func (s *Server) TaggerStreamEntities(in *pb.StreamTagsRequest, out pb.AgentSecu
 
 			responseEvents := make([]*pb.StreamTagsEvent, 0, len(events))
 			for _, event := range events {
-				e, err := pbutils.Tagger2PbEntityEvent(event)
+				e, err := proto.Tagger2PbEntityEvent(event)
 				if err != nil {
 					log.Warnf("can't convert tagger entity to protobuf: %s", err)
 					continue
@@ -127,7 +127,7 @@ func (s *Server) TaggerFetchEntity(ctx context.Context, in *pb.FetchEntityReques
 	}
 
 	entityID := fmt.Sprintf("%s://%s", in.Id.Prefix, in.Id.Uid)
-	cardinality, err := pbutils.Pb2TaggerCardinality(in.Cardinality)
+	cardinality, err := proto.Pb2TaggerCardinality(in.Cardinality)
 	if err != nil {
 		return nil, err
 	}
